Add npm command init & production
We have to make sure the api server is always on
and always recover when they crash. Using PM2 can
be a solution, so we've created two command to make
sure it's working.

Init command is used to install the pm2 cli on the
server before it run by the next command.

production command is used to run the app by passing
the entrypoint which is server.js to the pm2 and
adding an --name flag for better montioring the app